README: Fix inconsistent spacing of parameters
I noticed that the parameters in the POST functionality example have no spacing between them. All other examples of functions do have spacing around the parameters. This example function is therefore inconsistent the the rest of the example functions in the documentation.

The reason this actually matters to me is because I wanted to use this code in my project and had to change the spacing manually. I thought I would fix it and submit this pull request to help save the next person the trouble of correcting the spacing manually.